feat: PATCH API for default learning title
